Fix collision + Gravity particles

+++ src/gravityparticle.cpp
@@ -1,6 +1,6 @@
 #include "gravityparticle.hpp"
 
-GravityParticle::GravityParticle(int textureId, int time, double startVelX, double startVelY,
+GravityParticle::GravityParticle(int textureId, double time, double startVelX, double startVelY,
                                  double x, double y, double width, double height)
 	: Entity{textureId, x, y, width, height}, time{0}, timeTotal{time} {
 	velX = startVelX;
