+++ src/gravityparticle.hpp
@@ -1,11 +1,11 @@
 #pragma once
-#include <iostream>
+
 #include "entity.hpp"
 
 struct GravityParticle: public Entity
 {
 	GravityParticle() = default;
-	GravityParticle(int textureId, int time, double startVelX, double startVelY,
+	GravityParticle(int textureId, double time, double startVelX, double startVelY,
                     double x, double y, double width = 10, double height = 10);
 	~GravityParticle() = default;
 	
@@ -16,5 +16,5 @@ struct GravityParticle: public Entity
 	bool isDead() const;
 private:
     float time;
-	int timeTotal;
+    float timeTotal;
 };
